+++ Space-Simulator/Moons.hpp
@@ -15,20 +15,20 @@
 
 class Moon : protected SpaceObject {
 public:
-    Moon(std::string name_, int id_, long double mass_, long double longitude_, long double pericenter_, long double velocity_, bool retrograde_);
+    Moon(std::string name_, int id_, double mass_, double longitude_, double pericenter_, double velocity_, bool retrograde_);
     Moon() : Moon("Moon", rand(), 0, 0, 0, 0, false) {}
     
     std::string toString();
     
     bool equals(Moon other);
     
-    long double getLongitudeOfPericenter();
-    long double getPericenter();
+    double getLongitudeOfPericenter();
+    double getPericenter();
     bool isRetrograde();
     
 private:
-    long double longitudeOfPericenter; //meters (m)
-    long double pericenter; //meters (m)
+    double longitudeOfPericenter; //meters (m)
+    double pericenter; //meters (m)
     
     bool retrograde;
 };
