+++ Space-Simulator/StarSystemObjects.hpp
@@ -16,7 +16,7 @@
 class StarSystemObject : public SpaceObject {
 public:
     
-    StarSystemObject(std::string name_, int id_, long double mass_, long double longitude_, long double periastron_, long double velocity_, bool reversed_);
+    StarSystemObject(std::string name_, int id_, double mass_, double longitude_, double periastron_, double velocity_, bool reversed_);
     StarSystemObject() : StarSystemObject("Star System Object", rand(), 0, 0, 0, 0, false) {}
     StarSystemObject(const StarSystemObject &obj);
     ~StarSystemObject();
@@ -25,14 +25,14 @@ public:
     
     bool equals(StarSystemObject other);
     
-    long double getLongitudeOfPeriastron();
-    long double getPeriastron();
+    double getLongitudeOfPeriastron();
+    double getPeriastron();
     bool isReversed();
     
 protected:
     
-    long double longitudeOfPeriastron; //meters (m)
-    long double periastron; //meters (m)
+    double longitudeOfPeriastron; //meters (m)
+    double periastron; //meters (m)
     
     bool reversed;
